backend/requirements.txt
python-dotenv
flask 
flask_cors 
joblibp
supabase
pandas
numpy 
xgboost
scikit-learn
psycopg2
opencv-python
pillow
rembg
onnxruntime
tensorflow
imbalanced-learn
matplotlib
imblearn
